pyproject.toml
[project]
name = "recaptchav3"
version = "0.1.0"
description = "Quick recaptchav3 integration with DRF"
authors = [
    {name = "Esteban Chacon Martin",email = "etherbeing99@proton.me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
